(svn r16735) -Codechange: had_vehicle_of_type only stored 7 bits, all less that bit 7... but it was a word wasting space etc.

+++ src/station_base.h
@@ -110,7 +110,7 @@ public:
 
 	ViewportSign sign;
 
-	uint16 had_vehicle_of_type;
+	byte had_vehicle_of_type;
 
 	byte time_since_load;
 	byte time_since_unload;
